% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_batches.R
\name{geocode_batches}
\alias{geocode_batches}
\title{Geocode in batches}
\usage{
geocode_batches(
  .,
  batch_size,
  wait_time,
  n_results = 1,
  unwrap = FALSE,
  geonames_username
)
}
\arguments{
\item{.}{data frame}

\item{batch_size}{size of each batch to geocode}

\item{wait_time}{in seconds between batches}

\item{n_results}{same as geocode}

\item{unwrap}{as in geocode}

\item{geonames_username}{as in geocode}
}
\value{
df but geocoded in batches
}
\description{
Geocode in batches
}
\examples{
Coming soon
}
