% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emdat.R
\name{read_emdat}
\alias{read_emdat}
\title{Reads Excel Files obtained from EM-DAT Database}
\usage{
read_emdat(path_to_file, file_data = TRUE)
}
\arguments{
\item{path_to_file}{A String, the Path to the file downloaded.}

\item{file_data}{A Boolean, Do you want information about the file and how it was created?}
}
\value{
Returns a list containing one or two tibbles, one for the Disaster Data, and one for File Metadata.
}
\description{
Reads Excel files downloaded from the EMDAT Database linked \href{https://public.emdat.be/}{here}
}
\examples{
\dontrun{
read_emdat(path_to_file = "~/dummy", file_data = TRUE)
}
}
